Move most code of main.rs in lib.rs
