+++ src/commands/autocomplete.rs
@@ -2,7 +2,7 @@ use std::io::Write;
 
 use structopt::{clap::Shell, StructOpt};
 
-use crate::{Opt, Result, EXECUTABLE_NAME};
+use crate::{constants::EXECUTABLE_NAME, Opt, Result};
 
 pub fn run<W>(shell: Shell, buf: &mut W) -> Result<()>
 where
